requirements.txt
fastapi
uvicorn
easyocr
rapidfuzz
python-dotenv
supabase
